+++ tests/precision_unused_with_clause_finding.adb
@@ -1,0 +1,7 @@
+with Ada.Text_IO;
+with Ada.Strings.Unbounded;
+
+procedure Precision_Unused_With_Clause_Finding is
+begin
+   Ada.Text_IO.Put_Line ("hi");
+end Precision_Unused_With_Clause_Finding;